broken on crop step
IntroI've been working on AZ_hd_image_splitter.m on Viz2. This is a high definition version of a script that originally took in one bin10 ~6GB ROI of many nuclei and broke them up into crops of a single nucleus each. What this new script is doing is taking those same crops, from the bin10 data, and blow it up by 10x to crop the same regions out of the full resolution ~60 GB file of many nuclei.
I've left the script open on Viz2 and posted it to GitHub. Please note that it has some dependencies, and of course relies on large image files. And the first section takes 30 minutes to finish running! It has already run in the current instance on Viz2, though. The second section takes a minute or two, and the third section is where all the action is. It usually takes a minute or two per nucleus, so I've made sure to put in break points to limit the for-loop. The script takes in one whole ROI of many nuclei and is supposed to break that set of nuclei up to save 2 TIFs per nucleus: an hd_crop and an hd_mask. Once that's working, Vignesh would also like two more of the same TIFs, but with the membrane eroded. I haven't gotten to that.

In lines 39-51, the script locates the nuclei and creates a crude mask, albeit cropped very well. Later, on lines 78-80, the script shoehorns in your much better mask (gokulMask) in place of the crude mask. Lastly, gokulMask is upscaled by a factor of 10. All of this seems to work well. 
ProblemMy skills reach their limit at upscaling the cropping coordinates (lines 67-72). I've tried adding the upscaling factor at various points in these lines, to no avail. The error has consistently been that lines 85 and 87 fail, i.e. rawCrop can't be matrix multiplied by finalMask because rawCrop ends up several voxels smaller than finalMask. For example, finalMask would be 1620x1650x1590 uint8, while rawCrop would be 1611x1641x1581 uint8.
Notes On Attempted SolutionsChecked the script AZ_GU_cavitySearch.m for possible relevance to getting perfect crop coordinates. This was a dead end because the script only served to create improved masks from already cropped raw data.Read the documentation carefully on ind2sub(). Carefully and repeatedly thought through what lines 67-69 were doing to make sure I understood. Took breaks and came back to this problem over and over, with no clear improvement. Spoke with Vignesh.Moved the upscaling factor, *10, around within and among the parentheses on lines 67-72 to no avail. Experimented with transforming the data types of finalMask to and from logical and uint8 (see lines 80, 82, and 88). This may be a key part that I'm screwing up.Added and subtracted 4 or 5 voxels in each dimension (lines 70-72) to jerry rig the dimensions of rawCrop to fit the dimensions of finalMask, e.g. from 1611x1641x1581 uint8 to  1620x1650x1590 uint8. This is clearly not a repeatable solution. While it did prevent the dimension mismatch error, it manifested a new problem:A New ProblemEven when rawCrop and finalMask had matching dimensions, the HD_crop output file was a binary mask instead of a rich, HD segmentation of the raw data